server/requirements.txt
fastapi
pymongo
python-dotenv
passlib[bcrypt]
python-jose
uvicorn
